refactor(daemon): split request operation modules
